Replays and AI improvements
+ Hosts can now choose to make their games private.
++ Private games are not seen in the lobby but CAN still be joined
++ When the host makes the game private he will be shown a link which he
can use to invite other players. That share link is basically the one you
WOULD get from the lobby

+ A new feature was added. Players can now rewatch a game via the observer
mode which will provide them with a breakdown of what had happened.
++ This will allow players too see the highlighted danger lines which they
might have missed and understand why they lost or why the other player
played the way he did.
++ After Rewatching a game you may record it but it will not be visible
(will be private)! If you wanted a record you should have clicked
"Record"!

+ Leaving a game is improved and it will not destroy the game anymore.
This will allow players to rewatch and record the game even though their
opponent has left
++ When a player leaves only his session_id is cleared, unlike before
where the entire game entry was deleted
++ Recording and Rewatching when the opponent leaves is possible but the
links are NOT provided. If you know how to achieve this manually, fine,
otherwise we stick with the idea that the leaver does not need to be
shamed further with a record

+ The AI has been slightly changed.
++ The "Aggressive::Combo" Section during evaluation is improved to
consider dangerous5
++ The start of a game will now trigger all of the AIs into playing,
instead of just the first sequent. This, however, is rather useless
because they still don't play as they should in the game.
+- The AI is not made to play versus multiple opponents. If you still want
to play like that you can fix the bugs associated by calling /SCU/?ai
PARALLEL to your game tab.